move learner profile to core
