+++ src/main/webapp/app/core/user/settings/learner-profile/course-learner-profile.component.spec.ts
@@ -1,11 +1,11 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { CourseLearnerProfileComponent } from 'app/core/user/settings/learner-profile/course-learner-profile.component';
 import { CourseManagementService } from 'app/core/course/manage/services/course-management.service';
-import { LearnerProfileApiService } from 'app/learner-profile/service/learner-profile-api.service';
+import { LearnerProfileApiService } from 'app/core/learner-profile/service/learner-profile-api.service';
 import { Course } from 'app/core/course/shared/entities/course.model';
 import { of } from 'rxjs';
 import { HttpResponse, provideHttpClient } from '@angular/common/http';
-import { CourseLearnerProfileDTO } from 'app/learner-profile/shared/entities/learner-profile.model';
+import { CourseLearnerProfileDTO } from 'app/core/learner-profile/shared/entities/learner-profile.model';
 import { MockProvider } from 'ng-mocks';
 import { AlertService } from 'app/shared/service/alert.service';
 import { MockAlertService } from 'test/helpers/mocks/service/mock-alert.service';
