+++ src/main/webapp/app/shared/segmented-toggle/segmented-toggle.component.ts
@@ -1,6 +1,6 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { NgClass } from '@angular/common';
-import { CourseLearnerProfileLevel } from 'app/learner-profile/shared/entities/learner-profile-options.model';
+import { CourseLearnerProfileLevel } from 'app/core/learner-profile/shared/entities/learner-profile-options.model';
 
 @Component({
     selector: 'jhi-segmented-toggle',
